Created new services directory to contain business logic and moved portfolio analyzer module there
